Update License.txt
Had to update the license from Apache to Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International Public License (CC BY NC SA) because the maker of the game have the same and the license requires it. There is no plaintext version online so I had to number all of the sectors and list myself.